packages tests and main split, moved to main: api, data, locators
