+++ src/main/java/com/softserve/edu/greencity/ui/data/Goal.java
