+++ src/main/java/com/softserve/edu/greencity/ui/data/econews/NewsData.java
