+++ src/main/java/com/softserve/edu/greencity/ui/locators/EcoNewsPageLocator.java
