+++ src/main/java/com/softserve/edu/greencity/ui/locators/SingleNewsPageLocators.java
