+++ src/test/java/com/softserve/edu/greencity/ui/tests/runner/RemoteSkipTestAnalyzer.java
@@ -1,11 +1,8 @@
 package com.softserve.edu.greencity.ui.tests.runner;
 
-import com.softserve.edu.greencity.ui.tools.api.google.sheets.ValueProvider;
-import org.apache.commons.exec.environment.EnvironmentUtils;
+import com.softserve.edu.greencity.ui.api.google.sheets.ValueProvider;
 import org.testng.*;
-import org.testng.annotations.ITestAnnotation;
 
-import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 
 public class RemoteSkipTestAnalyzer implements IInvokedMethodListener {
